Added extensive (but not yet complete) support for the search JTextField.variant. Still need to add support for the actions. Setting JTextField.Search.FindPopup to a PopupMenu draws the popup icon, but it currently has no effect.
We also added code to support a JTextField.Search.PlaceholderText value, which will be drawn in the field when there is no user text present and the field does not have focus. This will be drawn in gray.

Pressing the "cancel" button deletes the text, but currently has no other effect.